refactor: moved http handlers to a separate handlers package and simplified main.go
